requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
black==26.1.0
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycodestyle==2.14.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
python-slugify==8.0.4
pytokens==0.4.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
